third_party/WebKit/LayoutTests/editing/inserting/insert-html-into-text-field.html

<head>
<style>
#xtest {
    x-border-style: none;
    box-sizing: border-box;
    font-size: 15px;
    height: 30px;
    padding: 15px;
}
</style>
</head>
<body>
<div id="content">
<p id="description"></p>
Steps:
<ol>
    <li>Set caret after "foo" in below text field.</li>
    <li>Hit Ctrl+V to paste text.</li>
    <li>You should see pasted text in text field.</li>
</ol>
Text Field:<input id="test" value="foo">
</div>
<script src="../../resources/js-test.js"></script>
<script>
description('InsertHTML should not be affected by CSS style');
var textField = document.getElementById('test');
textField.focus();
textField.setSelectionRange(textField.value.length, textField.value.length);
document.execCommand('InsertHTML', false, 'bar');
shouldBeEqualToString('textField.value', 'foobar');

if (window.testRunner)
    document.getElementById('content').outerHTML = '';
</script>
</body>
