third_party/blink/web_tests/editing/inserting/insert-html-into-text-field-expected.txt

PASS textField.value is "foobar"
PASS successfullyParsed is true

TEST COMPLETE
